pyproject.toml
[project]
name = "akilan"
version = "0.0.3"
description = "Adaptive Knowledge Ingestion Layer for Analytics Nodes – PyMuPDF wrapper for AI-ready document processing"
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }

authors = [
  { name = "Akilan"}
]

dependencies = [
  "loguru>=0.7.3",
  "pillow>=12.1.1",
  "pymupdf>=1.27",
  "rich>=14.3.3",
]

keywords = [
  "pdf",
  "pymupdf",
  "document-processing",
  "ai",
  "text-extraction"
]

classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Text Processing",
  "Topic :: Scientific/Engineering :: Artificial Intelligence"
]

[project.urls]
Homepage = "https://github.com/Akilankm/akilan"
Repository = "https://github.com/Akilankm/akilan"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/akilan"]

[dependency-groups]
dev = [
    "build>=1.4.0",
    "twine>=6.2.0",
]
